Address review feedback on Tokio metrics
Add the GaugeExt::record_if_nonzero helper from the original snippet and
use it instead of inlining the nonzero checks.

Rename QUEUE_DEPTH_REPORT_THRESHOLD to SATURATION_REPORT_THRESHOLD, since
it also gates a thread count rather than a queue depth, and document why
low values are not worth reporting.

Drop the bucket bounds check in the poll histogram loop. The runtime is
built with the default poll-time histogram, which has exactly
NUM_REPORTED_POLL_HISTOGRAM_BUCKETS buckets, so the index cannot go out of
range.

Move spawn_tokio_metrics into metrics.rs and have it build its own
TokioMetrics.
